Factor out the ExecuteI block
